lib(refactor[sanitize]): extract sanitize module
why: sanitization lived in publisher.py, so the writer could not
  reuse it without a circular import; the fix needs the writer to
  sanitize after transforms, and the pass is worth exposing on
  its own.
what:
- Move sanitize_doctree, _uri_is_allowed, and node helpers to
  lib/sanitize.py; publisher re-imports sanitize_doctree
- Add SanitizeTransform for custom docutils pipelines
- Point the security tests at the new module
